Question 1.java
Write a Java program where the current year and the year in which an employee joined the organization are entered through the keyboard. If the number of years for which the employee has served the organization is greater than 5, a bonus of Rs. 5000/- is given to the employee. If the years of service are between 3 and 5 (inclusive), a bonus of Rs. 3000/- is given. If the years of service are less than 3, then the program should print a message indicating that no bonus is awarded.