Question 2.java
A library charges a fine for every book returned late. For the first 7 days, the fine is 50 paise, for 8-14 days the fine is one rupee, and above 14 days fine is 5 rupees. If you return the book after 21 days, your membership will be canceled. Write a program to accept the number of days the member is late to return the book and display the fine or the appropriate message.